+++ src/WordPress/Bridge.php
@@ -20,6 +20,8 @@ class Bridge {
 
 	protected $userClass = '\FatPanda\Illuminate\WordPress\Models\User';
 
+	static protected $databaseManager;
+
 	/**
 	 * Create a new Bridge.
 	 * @param String namespace
@@ -30,6 +32,7 @@ class Bridge {
 	function __construct($namespace, $laravelAppBoostrapFilePath = null, $userClass = null, $pluginMainFilePath = null)
 	{
 		$this->namespace = $namespace;
+		
 		if ($pluginMainFilePath) {
 			$this->setPluginMainFilePath($pluginMainFilePath);
 		}
@@ -170,29 +173,33 @@ class Bridge {
 
 	/**
 	 * Enable Eloquent for manipulating data in the database.
-	 * If you have Bridged with a Laravel or Lumen bootstrap file,
-	 * Laravel will handle this for you. But if you're not using
-	 * Laravel, you can still use Illuminate/Database.
-	 * @param mixed Connection arguments; this is backfilled with
-	 * WordPress' global defaults
-	 * @param bool Set this connection as the global connection; defaults to true
-	 * @return \Illuminate\Database\Capsule\Manager
-	 * @see http://github.com/illuminate/database
+	 * @deprecated since 1.0.7
+	 * @return Illuminate\Database\Capsule\Manager
+	 * @see Bridge::bootEloquent()
 	 */
-	function enableEloquent($connection = '', $setAsGlobal = true)
+	function enableEloquent()
+	{
+		$this->bootEloquent();
+	}
+
+	/**
+	 * Enable Eloquent for manipulating data in the database.
+	 * @return Illuminate\Database\Capsule\Manager
+	 */
+	function bootEloquent()
 	{
 		global $table_prefix;
 
 		// if this Bridge is configured with a Laravel App,
 		// then we let the app setup Eloquent
 		if ($this->hasLaravelApp()) {
+			$this->bootstrap();
 			return true;
 		}
 
-		$capsule = new \Illuminate\Database\Capsule\Manager;
+		$capsule = static::getDatabaseManager();
 
-
-		$capsule->addConnection(wp_parse_args($connection, [
+		$this->addConnection([
 			'driver' => 'mysql',
 			'host' => DB_HOST,
 			'database' => DB_NAME,
@@ -201,15 +208,47 @@ class Bridge {
 			'charset' => DB_CHARSET,
 			'collation' => DB_COLLATE,
 			'prefix' => $table_prefix,
-		]));
-
-		if ($setAsGlobal) {
-			$capsule->setAsGlobal();
-		}
+		], 'default');
 
 		$capsule->bootEloquent();
 
 		return $capsule;
+	}
+
+	/**
+	 * @param mixed Connection arguments
+	 * @param String The name to identify this connection by; defaults to 'default'
+	 * @return Illuminate\Database\Capsule\Manager
+	 * @see http://github.com/illuminate/database
+	 */
+	function addConnection($connection = '', $name = 'default')
+	{
+		$capsule = static::getDatabaseManager();
+
+		try {
+			if ($capsule->getConnection($name)) {
+				return $capsule;
+			}
+		} catch (\InvalidArgumentException $e) {
+			// ignore; it's just because the connection doesn't exist
+		}
+
+		$capsule->addConnection(wp_parse_args($connection), $name);
+		
+		return $capsule;
+	}
+
+	/**
+	 * Setup the database manager.
+	 * @return Illuminate\Database\Capsule\Manager
+	 */
+	static function getDatabaseManager()
+	{
+		if (empty(static::$databaseManager)) {
+			static::$databaseManager = new \Illuminate\Database\Capsule\Manager;
+			static::$databaseManager->setAsGlobal();
+		}
+		return static::$databaseManager;
 	}
 
 	/**