+++ src/WordPress/Plugin.php
@@ -37,6 +37,8 @@ abstract class Plugin {
 
 	protected $routerVersion;
 
+	protected $reflection;
+
 	function __construct($mainFile)
 	{
 		if (!file_exists($mainFile)) {
@@ -44,6 +46,77 @@ abstract class Plugin {
 		}
 
 		$this->mainFile = $mainFile;
+		
+		$this->bindActionsAndFilters();
+	}
+
+	/**
+	 * Using reflection, put together a list of all the action and filter hooks
+	 * defined by this class, and then setup bindings for them.
+	 * Action hooks begin with the prefix "on" and filter hooks begin with the
+	 * prefix "filter". Additionally, look for the @priority doc comment, and
+	 * use that to configure the priority loading order for the hook. Finally, count
+	 * the number of parameters in the method signature, and use that to control
+	 * the number of arguments that should be passed to the hook when it is invoked.
+	 */
+	protected function bindActionsAndFilters()
+	{
+		// setup activation and deactivation hooks
+		$this->bindActivationAndDeactivationHooks();
+
+		// reflect on the contents of this class
+		$this->reflection = new \ReflectionClass($this);
+
+		// get a list of all the methods on this class
+		$methods = $this->reflection->getMethods(\ReflectionMethod::IS_PUBLIC);
+
+		// look for candidates for actions or filter hooks
+		foreach($methods as $method) {
+			// skip activation/deactivation hooks (handled above)
+			if ($method->getName() === 'onActivate' || $method->getName() === 'onDeactivate') {
+				continue;
+			}
+
+			// look in codedoc for @priority
+			$priority = 10;
+			$docComment = $method->getDocComment();
+			if ($docComment !== false) {
+				if (preg_match('/@priority\s+(\d+)/', $docComment, $matches)) {
+					$priority = (int) $matches[1];
+				}
+			}
+
+			// setup some internal event handlers
+			add_action('plugins_loaded', [ $this, 'finalOnPluginsLoaded' ], 9);
+			add_action('init', [ $this, 'finalOnInit' ], 9);
+			
+			// action methods begin with "on"
+			if ('on' === strtolower(substr($method->getName(), 0, 2))) {
+				$action = trim(strtolower(preg_replace('/(?<!\ )[A-Z]/', '_$0', substr($method->getName(), 2))), '_');
+				$parameterCount = $method->getNumberOfParameters();
+				add_action($action, [ $this, $method->getName() ], $priority, $parameterCount);
+			
+			// filter methods begin with "filter"
+			} else if ('filter' === strtolower(substr($method->getName(), 0, 6))) {
+				$filter = trim(strtolower(preg_replace('/(?<!\ )[A-Z]/', '_$0', substr($method->getName(), 6))), '_');
+				$parameterCount = $method->getNumberOfParameters();
+				add_action($filter, [ $this, $method->getName() ], $priority, $parameterCount);
+			}
+		}
+	}
+
+	/**
+	 * Load plugin meta data, finish configuring various features, including
+	 * the REST router and text translation.
+	 * @see https://codex.wordpress.org/Plugin_API/Action_Reference/plugins_loaded
+	 */
+	final function finalOnPluginsLoaded()
+	{
+		// if we don't have the get_plugin_data() function, load it
+		if (!function_exists('get_plugin_data')) {
+			require_once ABSPATH.'wp-admin/includes/plugin.php';
+		}
+
 		$this->pluginData = get_plugin_data($this->mainFile);
 		
 		foreach($this->pluginData as $key => $value) {
@@ -51,6 +124,56 @@ abstract class Plugin {
 			$this->{$propertyName} = $value;
 		}
 
+		$this->loadRouterAndRoutes();
+
+		$this->loadTextDomain();
+	}
+
+	/**
+	 * @see https://codex.wordpress.org/Plugin_API/Action_Reference/init
+	 */
+	final function finalOnInit()
+	{
+		$this->registerCustomDataTypes();		
+	}
+
+	/**
+	 * Allow for methods of Bridge to be invoked on Plugin
+	 */
+	function __call($name, $args)
+	{
+		$callback = [ $this->router->bridge, $name ];
+		if (is_callable($callback)) {
+			return call_user_func_array($callback, $args);
+		} else {
+			throw new \BadMethodCallException($name);
+		}
+	}
+
+	protected function bindActivationAndDeactivationHooks()
+	{
+		register_activation_hook($this->mainFile, [ $this, 'onActivate' ]);
+		register_deactivation_hook($this->mainFile, [ $this, 'onDeactivate' ]);
+	}
+
+	protected function registerCustomDataTypes()
+	{
+		if (!empty($this->registeredDataTypes)) {
+			$this->router->bridge->bootEloquent();
+
+			foreach($this->registeredDataTypes as $class) {
+				call_user_func($class . '::register');
+			}
+		}
+	}
+
+	protected function loadTextDomain()
+	{
+		load_plugin_textdomain( $this->textDomain, false, dirname( plugin_basename($this->mainFile) ) . rtrim($this->domainPath, '/') . '/' );
+	}
+
+	protected function loadRouterAndRoutes()
+	{
 		$this->setRouterNamespace( Str::slug($this->pluginName) );
 
 		// just use the major version number
@@ -61,15 +184,11 @@ abstract class Plugin {
 
 		$this->setRouterVersion($routerVersion);
 
-		register_activation_hook($this->mainFile, [ $this, 'activate' ]);
-		register_deactivation_hook($this->mainFile, [ $this, 'deactivate' ]);
-		add_action('plugins_loaded', [ $this, 'loadTextDomain' ]);
-	}
+		$this->router = $router = new Router($this->routerNamespace, $this->routerVersion);
 
-	function loadTextDomain()
-	{
-		load_plugin_textdomain( $this->textDomain, false, 
-			dirname( plugin_basename($this->mainFile) ) . rtrim($this->domainPath, '/') . '/' );
+		$plugin = $this;
+
+		require $this->getPluginBasePath() . '/src/Http/routes.php';
 	}
 
 	function getPluginBasePath()
@@ -80,19 +199,6 @@ abstract class Plugin {
 	function getPluginData()
 	{
 		return $this->pluginData;
-	}
-
-	function make()
-	{
-		$this->router = $router = new Router($this->routerNamespace, $this->routerVersion);
-		
-		if (!empty($this->registeredDataTypes)) {
-
-		}
-
-		$plugin = $this;
-
-		require $this->getPluginBasePath() . '/src/Http/routes.php';
 	}
 
 	function setRouterNamespace($namespace)
@@ -113,13 +219,18 @@ abstract class Plugin {
 		return $this;
 	}
 
-	abstract function activate();
+	abstract function onActivate();
 
-	abstract function deactivate();
+	abstract function onDeactivate();
 
 	function register($customDataType)
 	{
-		$this->registeredDataTypes[] = $customDataType;
+		$this->registeredDataTypes[(string) $customDataType] = $customDataType;
+	}
+
+	function unregister($customDataType)
+	{
+		unset($this->registeredDataTypes[(string) $customDataType]);
 	}
 
 }