+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## v1.0.6
+
+* Fixed bugs in CustomPostType
+* Introduced CustomTaxonomy
+* Established naming convention for Plugin hooks: on{ActionName} for actions, and filter{FilterName} for filters, with support for @priority codedoc
+* Deprecated Bridge::enableEloquent in favor of Bridge::bootEloquent (matches Laravel method name for the same)
+* Refactored the way database connections are created, hopefully ensuring that there are never more than two opened for one WordPress request
+
 ## v1.0.5
 
 * Refactored Plugin baseclass out of scaffolding project and into this one